Old Stripy result incompatibility bug
This Pull request built on top of changes made in PR_4

There was a bug when making stripy reports from old stripy results where some essential fields were missing, in this case it was FullyRepeatedReadSources. 

A parsing function was made to ensure that all essential fields existed at the Make stripy reports stage. Furthermore, the index page stage is forced every time the job is run as it produces and overwrites the index page once per project.

make_stripy_reports.py contains the default report schema. Importantly, the Allele fields are essential but not parsed here as we want the report to fail if data for alleles does not exist.

Lastly, cool cosmetic html magic was done too, hopefully this issue is now closed!